+++ src/index.ts
@@ -12,6 +12,7 @@ export * from './checkbox';
 export * from './input-number';
 export * from './config-provider';
 export * from './steps';
+export * from './sticky-tool';
 export * from './message';
 export * from './table';
 export * from './tag';