+++ src/sticky-tool/defaultProps.ts
@@ -2,13 +2,13 @@
  * 该文件为脚本自动生成文件，请勿随意修改。如需修改请联系 PMC
  * */
 
-import { TdStickyItemProps, TdStickyToolProps } from './type';
-
-export const stickyItemDefaultProps: TdStickyItemProps = { trigger: 'hover' };
+import { TdStickyToolProps, TdStickyItemProps } from './type';
 
 export const stickyToolDefaultProps: TdStickyToolProps = {
-  collapsed: false,
-  draggable: false,
   list: [],
   placement: 'right-bottom',
+  shape: 'square',
+  type: 'normal',
 };
+
+export const stickyItemDefaultProps: TdStickyItemProps = { trigger: 'hover' };